SAK-30032 Allow Student to attach files during Peer Review
This allows students to add one or more attachments to their reviews when using the Peer Review feature in Assignments.
There is an additional table that is needed for this feature. The conversion scripts for this table can be found in
sakai_11_mysql_conversion.sql or sakai_11_oracle_conversion.sql